# apps/docs/public/openapi-books.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Syncero Books API",
    "version": "1.0.0",
    "description": "API privada do **Syncero Books** — app para contadores.\n\n### Autenticação\n- **Rotas internas:** Bearer JWT emitido pelo Supabase (requer `Origin` de domínio Syncero)\n- **Rotas externas:** `x-api-key` (API Key do sistema) ou Bearer JWT\n\n### Restrição de Origem (Origin Guard)\nRotas internas verificam o header `Origin` em produção.\nApenas origens em `ALLOWED_ORIGINS` são aceitas — rotas externas não têm essa restrição.\n\n### Rotas externas (aceitam x-api-key)\n- `GET /api/transactions` — listar lançamentos de empresa externa\n- `POST /api/transactions` — criar lançamento para empresa externa\n- `POST /api/journal-entries` — classificar lançamento (criar partida dobrada)\n\n### Autorização interna\n- Empresas Flow: `accountant_companies.status = 'accepted'`\n- Empresas externas: `external_companies.accountant_id = userId`"
  },
  "servers": [
    {
      "url": "http://localhost:5175",
      "description": "Desenvolvimento local"
    },
    {
      "url": "https://syncero-books.vercel.app",
      "description": "Produção"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKey": []
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API Key gerada pelo sistema (GET/POST /api-keys). Aceita nas rotas externas como alternativa ao Bearer JWT."
      },
      "originGuard": {
        "type": "apiKey",
        "in": "header",
        "name": "Origin",
        "description": "Header de origem verificado nas rotas internas. Em produção, apenas origens Syncero são aceitas."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "TransactionType": {
        "type": "string",
        "enum": [
          "income",
          "expense"
        ]
      },
      "TransactionNature": {
        "type": "string",
        "enum": [
          "sale_service",
          "loan_received",
          "capital_contribution",
          "operational_expense",
          "product_cost",
          "asset_purchase",
          "debt_payment",
          "owner_withdrawal"
        ]
      },
...
      "Transaction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "company_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "ext_company_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "description": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "type": {
            "$ref": "#/components/schemas/TransactionType"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "is_paid": {
            "type": "boolean"
          },
          "paid_at": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "nature": {
            "$ref": "#/components/schemas/TransactionNature"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "is_classified": {
            "type": "boolean",
            "description": "Adicionado pelo servidor — indica se há lançamento contábil vinculado"
          },
          "journal_entry_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PaginatedTransactions": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Transaction"
            }
          },
          "count": {
            "type": "integer"
          }
        }
      },
...
    },
...
  },
  "paths": {
...
    "/api/transactions/counterparts": {
      "get": {
        "tags": [
          "transactions-external"
        ],
        "summary": "Listar counterparts distintos de empresa externa",
        "description": "Retorna lista ordenada de valores únicos de `counterpart` para autocomplete.\n\n**Rota externa** — aceita `x-api-key` além de Bearer JWT.",
        "parameters": [
          {
            "name": "ext_company_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array de strings únicas, ordenado alfabeticamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "type": "string" }
                }
              }
            }
          },
          "400": {
            "description": "extCompanyId obrigatório"
          },
          "403": {
            "description": "Sem acesso à empresa"
          }
        }
      }
    },
    "/api/transactions": {
      "get": {
        "tags": [
          "transactions-external"
        ],
        "summary": "Listar lançamentos com status de classificação",
        "description": "Aceita `companyId` (empresa Flow, somente leitura) ou `extCompanyId` (empresa externa, CRUD completo).\nRetorna `is_classified` e `journal_entry_id` adicionados pelo servidor.\n\n\n**Rota externa** — aceita `x-api-key` além de Bearer JWT. Sem restrição de Origin.",
        "parameters": [
          {
            "name": "company_id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Empresa Flow (exclusivo com extCompanyId)"
          },
          {
            "name": "ext_company_id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Empresa externa (exclusivo com companyId)"
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/TransactionType"
            }
          },
          {
            "name": "is_paid",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 1000
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Página de lançamentos com status de classificação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedTransactions"
                }
              }
            }
          }
        },
        "x-access": "external",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ]
      },
      "post": {
        "tags": [
          "transactions-external"
        ],
        "summary": "Criar lançamento para empresa externa",
        "description": "Exclusivo para empresas externas (`extCompanyId`). Empresas Flow são somente leitura no Books.\n\n**Rota externa** — aceita `x-api-key` além de Bearer JWT. Sem restrição de Origin.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "ext_company_id",
                  "description",
                  "amount",
                  "type",
                  "date",
                  "is_paid",
                  "nature",
                  "counterpart"
                ],
                "properties": {
                  "ext_company_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "description": {
                    "type": "string",
                    "example": "Aluguel de escritório"
                  },
                  "amount": {
                    "type": "number",
                    "minimum": 0.01
                  },
                  "type": {
                    "$ref": "#/components/schemas/TransactionType"
                  },
                  "date": {
                    "type": "string",
                    "format": "date"
                  },
                  "is_paid": {
                    "type": "boolean"
                  },
                  "paid_at": {
                    "type": "string",
                    "format": "date",
                    "nullable": true
                  },
                  "nature": {
                    "$ref": "#/components/schemas/TransactionNature",
                    "description": "Natureza contábil. Alimenta sugestões automáticas de classificação."
                  },
                  "counterpart": {
                    "type": "string",
                    "description": "Quem enviou (receita) ou recebeu (despesa) o valor",
                    "example": "Cliente XYZ Ltda"
                  },
                  "notes": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Lançamento criado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          }
        },
        "x-access": "external",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ]
      }
    },
    "/api/transactions/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "tags": [
          "transactions-internal"
        ],
        "summary": "Obter lançamento com detalhes expandidos",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Transaction"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "creator_name": {
                          "type": "string",
                          "nullable": true
                        },
                        "payment_registrar_name": {
                          "type": "string",
                          "nullable": true
                        },
                        "categories": {
                          "type": "object",
                          "nullable": true
                        },
                        "banks": {
                          "type": "object",
                          "nullable": true
                        },
                        "contacts": {
                          "type": "object",
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado ou origem não autorizada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Não encontrado"
          }
        },
        "x-access": "internal",
        "description": "\n\n**Nota de segurança:** Retorna 404 (não 403) quando o recurso existe mas o usuário não tem acesso, para evitar oracle de existência."
      },
      "patch": {
        "tags": [
          "transactions-external"
        ],
        "summary": "Atualizar lançamento de empresa externa",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Campos editáveis. **Ignorados mesmo se enviados:** ext_company_id, company_id, created_by, created_at, updated_at.",
                "x-non-writable": [
                  "ext_company_id",
                  "company_id",
                  "created_by",
                  "created_at",
                  "updated_at"
                ],
                "properties": {
                  "description": {
                    "type": "string"
                  },
                  "amount": {
                    "type": "number",
                    "minimum": 0.01
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "income",
                      "expense"
                    ]
                  },
                  "date": {
                    "type": "string",
                    "format": "date"
                  },
                  "is_paid": {
                    "type": "boolean"
                  },
                  "paid_at": {
                    "type": "string",
                    "format": "date",
                    "nullable": true
                  },
                  "nature": {
                    "$ref": "#/components/schemas/TransactionNature",
                    "description": "Não pode ser removido após definido"
                  },
                  "counterpart": {
                    "type": "string",
                    "description": "Quem enviou (receita) ou recebeu (despesa) — não pode ser removido"
                  },
                  "notes": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          }
        },
        "x-access": "external",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "description": "Atualiza lançamento de empresa externa. **Rota externa** — aceita x-api-key."
      },
      "delete": {
        "tags": [
          "transactions-external"
        ],
        "summary": "Excluir lançamento de empresa externa",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        },
        "x-access": "external",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ]
      }
    },
...
  }
}